requirements.txt
fastapi
strawberry-graphql[fastapi]
sqlalchemy
aiosqlite
uvicorn